packages/collector/pyproject.toml
[project]
name = "monctl-collector"
version = "2.0.0"
description = "MonCTL distributed collector — cache-node, poll-worker, forwarder"
requires-python = ">=3.11"
dependencies = [
    # gRPC peer communication
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    # Async HTTP client (central API)
    "aiohttp>=3.9.0",
    # Async SQLite (local cache)
    "aiosqlite>=0.19.0",
    # YAML config
    "pyyaml>=6.0",
    # Credential encryption (Fernet)
    "cryptography>=42.0.0",
    # Logging
    "structlog>=24.1",
    # TUI
    "textual>=1.0.0",
    # Sync HTTP client (for TUI workers)
    "httpx>=0.27.0",
]

[project.scripts]
monctl-cache-node    = "monctl_collector.entrypoints.cache_node:main"
monctl-poll-worker   = "monctl_collector.entrypoints.poll_worker:main"
monctl-forwarder     = "monctl_collector.entrypoints.forwarder_main:main"
monctl-setup         = "monctl_collector.tui.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monctl_collector"]
